Create single element in a sorted array
In this question, we check whether there is an element present which occur only one time. we check for first and last element, middle element by checking previous and next element(duplicates). Also if mid is even and element is not equal to next we change the start element and if odd and element is not equal to previous we change the start element only.