Create search insert position
this program checks whether the target element is there is the array, if there, it returns the position else it find the insert position where the target element should be inserted. This is all done by use of binary search. using mid and check where the target should be inserted either in the left portion or in the right.
 Time complexity: O(logn)
Auxiliary space: O(1)